Restore out of channel and some cleanup
